Resolve scrolling issues
This makes use of a relatively new part of CSS known as "Scroll Snap".
It's well-supported in all modern browsers, which should be fine for a
developer tool.

I'm not completely sure how much the Bootstrap Scrollspy offset change
is needed, but it seems to work well in its current state.